refactor and renaming the package. library and module imports are tested in the package_test.ipynb
