--- house_price_modeling_lin_yang/house_price_modeling_packaging/__init__.py
@@ -1,1 +1,0 @@
-from src.house_prices_prediction_lin import *
